Big bunch of improvements.
See http://dave.cheney.net/2014/01/30/avr11-profiling-on-the-atmega for the overview.

The next challenge is to remove the use of uint32_t's for unibus addresses. This will mean a union type to avoid the shift cost for 32bit quantities.

I should probably do the same for registers and other 16bit quantities like the PS word.